+++ ws-server/src/quackers_game/cracker_creator.rs
@@ -16,10 +16,10 @@ pub fn generate_random_cracker_data() -> CrackerData {
     // let colors_length: u64 = available_duck_colors.len() as u64;
 
     // let random_color_index: u64 = rng.gen_range(0..=(colors_length - 1));
-    // let random_quack_pitch: f64 = rng.gen_range(MIN_QUACK_MULTIPLIER..MAX_QUACK_MULTIPLIER);
+    // let random_quack_pitch: f32 = rng.gen_range(MIN_QUACK_MULTIPLIER..MAX_QUACK_MULTIPLIER);
 
-    let random_x_pos: u64 = rng.gen_range(MIN_X_POS..MAX_X_POS);
-    let random_y_pos: u64 = rng.gen_range(MIN_Y_POS..MAX_Y_POS);
+    let random_x_pos: f32 = rng.gen_range(MIN_X_POS..MAX_X_POS);
+    let random_y_pos: f32 = rng.gen_range(MIN_Y_POS..MAX_Y_POS);
 
     let random_point_value: u64 =
         BASE_CRACKER_POINT_VALUE + rng.gen_range(0..RANDOM_CRACKER_POINT_VALUE);
