+++ ws-server/src/quackers_game/types/player_join_msg.rs
@@ -13,8 +13,8 @@ pub struct NewJoinerData {
     pub player_uuid: String,
     pub player_friendly_name: String,
     pub color: String,
-    pub x_position: u64,
-    pub y_position: u64,
+    pub x_position: f32,
+    pub y_position: f32,
 }
 
 #[derive(Debug, Serialize)]
