+++ ws-server/src/quackers_game/types/player_move_msg.rs
@@ -3,8 +3,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize)]
 pub struct MoveRequestData {
-    pub x_direction: u64,
-    pub y_direction: u64,
+    pub x_direction: f32,
+    pub y_direction: f32,
 }
 
 #[derive(Debug, Serialize, Clone)]
@@ -12,10 +12,10 @@ pub struct MoveResponseData {
     pub player_uuid: String,
     pub player_friendly_name: String,
     pub color: String,
-    pub old_x_position: u64,
-    pub old_y_position: u64,
-    pub new_x_position: u64,
-    pub new_y_position: u64,
+    pub old_x_position: f32,
+    pub old_y_position: f32,
+    pub new_x_position: f32,
+    pub new_y_position: f32,
 }
 
 #[derive(Debug, Serialize)]
